Extract jsonSchema and results logic
