+++ packages/dmn-runner/package.json
@@ -25,11 +25,15 @@
     "test": "run-script-if --ignore-errors \"$(build-env tests.ignoreFailures)\" --bool \"$(build-env tests.run)\" --then \"jest --silent --verbose --passWithNoTests\""
   },
   "dependencies": {
+    "@kie-tools-core/react-hooks": "workspace:*",
     "@kie-tools/extended-services-api": "workspace:*",
     "@kie-tools/uniforms-patternfly": "workspace:*",
     "@patternfly/react-core": "^4.276.6",
     "@patternfly/react-icons": "^4.93.6",
     "ajv": "^6.12.6",
+    "deep-object-diff": "^1.1.9",
+    "json-refs": "^3.0.15",
+    "lodash": "^4.17.21",
     "moment": "^2.29.4",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
@@ -43,6 +47,7 @@
     "@kie-tools/root-env": "workspace:*",
     "@kie-tools/tsconfig": "workspace:*",
     "@types/jest": "^26.0.23",
+    "@types/lodash": "^4.14.168",
     "@types/react": "^17.0.6",
     "@types/react-dom": "^17.0.5",
     "jest": "^26.6.3",
